+++ lib/token.ts
@@ -33,8 +33,6 @@ import * as sdkCrypto from './crypto';
 import AuthSdkError from './errors/AuthSdkError';
 import OAuthError from './errors/OAuthError';
 import {
-  ACCESS_TOKEN_STORAGE_KEY,
-  ID_TOKEN_STORAGE_KEY,
   REDIRECT_OAUTH_PARAMS_NAME,
   REDIRECT_NONCE_COOKIE_NAME,
   REDIRECT_STATE_COOKIE_NAME
@@ -733,6 +731,24 @@ function renewToken(sdk: OktaAuth, token: Token): Promise<Token> {
   });
 }
 
+function renewTokens(sdk: OktaAuth, options: TokenParams): Promise<Tokens> {
+  options = Object.assign({
+    scopes: sdk.options.scopes,
+    authorizeUrl: sdk.options.authorizeUrl,
+    userinfoUrl: sdk.options.userinfoUrl,
+    issuer: sdk.options.issuer
+  }, options);
+
+  if (sdk.options.pkce) {
+    options.responseType = 'code';
+  } else {
+    options.responseType = ['token', 'id_token'];
+  }
+
+  return getWithoutPrompt(sdk, options)
+    .then(res => res.tokens);
+}
+
 function removeHash(sdk) {
   var nativeHistory = sdk.token.parseFromUrl._getHistory();
   var nativeDoc = sdk.token.parseFromUrl._getDocument();
@@ -817,22 +833,20 @@ function parseFromUrl(sdk, options: string | ParseFromUrlOptions): Promise<Token
     });
 }
 
-async function getUserInfo(sdk, accessTokenObject, idTokenObject): Promise<UserClaims> {
+async function getUserInfo(sdk, accessTokenObject: AccessToken, idTokenObject: IDToken): Promise<UserClaims> {
   // If token objects were not passed, attempt to read from the TokenManager
   if (!accessTokenObject) {
-    accessTokenObject = await sdk.tokenManager.get(ACCESS_TOKEN_STORAGE_KEY);
+    accessTokenObject = (await sdk.tokenManager.getTokens()).accessToken as AccessToken;
   }
   if (!idTokenObject) {
-    idTokenObject = await sdk.tokenManager.get(ID_TOKEN_STORAGE_KEY);
+    idTokenObject = (await sdk.tokenManager.getTokens()).idToken as IDToken;
   }
 
-  if (!accessTokenObject ||
-      (!isToken(accessTokenObject) && !accessTokenObject.accessToken && !accessTokenObject.userinfoUrl)) {
+  if (!accessTokenObject || !isAccessToken(accessTokenObject)) {
     return Promise.reject(new AuthSdkError('getUserInfo requires an access token object'));
   }
 
-  if (!idTokenObject ||
-    (!isToken(idTokenObject) && !idTokenObject.idToken)) {
+  if (!idTokenObject || !isIDToken(idTokenObject)) {
     return Promise.reject(new AuthSdkError('getUserInfo requires an ID token object'));
   }
 
@@ -879,6 +893,7 @@ export {
   parseFromUrl,
   decodeToken,
   renewToken,
+  renewTokens,
   getUserInfo,
   verifyToken,
   handleOAuthResponse,