+++ lib/types/OktaAuthOptions.ts
@@ -13,9 +13,12 @@
 import { StorageUtil } from './Storage';
 import { CookieOptions } from './Cookies';
 import { HttpRequestClient } from './http';
+import { OktaAuth } from './';
+import { AuthState } from './AuthState';
 
 export interface TokenManagerOptions {
   autoRenew?: boolean;
+  autoRemove?: boolean;
   secure?: boolean;
   storage?: string;
   storageKey?: string;
@@ -54,4 +57,6 @@ export interface OktaAuthOptions extends CustomUrls {
   transformErrorXHR?: (xhr: object) => any;
   headers?: object;
   maxClockSkew?: number;
+  transformAuthState?: (oktaAuth: OktaAuth, authState: AuthState) => Promise<AuthState>;
+  devMode?: boolean; 
 }