+++ test/spec/token.js
@@ -2961,6 +2961,119 @@ describe('token.renew', function() {
   });
 });
 
+describe('token.renewTokens', function() {
+  it('should return tokens', function() {
+    return oauthUtil.setupFrame({
+      oktaAuthArgs: {
+        pkce: false,
+        issuer: 'https://auth-js-test.okta.com',
+        clientId: 'NPSfOkH5eZrTy8PMDlvx',
+        redirectUri: 'https://example.com/redirect'
+      },
+      tokenRenewTokensArgs: [],
+      postMessageSrc: {
+        baseUri: 'https://auth-js-test.okta.com/oauth2/v1/authorize',
+        queryParams: {
+          'client_id': 'NPSfOkH5eZrTy8PMDlvx',
+          'redirect_uri': 'https://example.com/redirect',
+          'response_type': 'token id_token',
+          'response_mode': 'okta_post_message',
+          'state': oauthUtil.mockedState,
+          'nonce': oauthUtil.mockedNonce,
+          'scope': 'openid email',
+          'prompt': 'none'
+        }
+      },
+      time: 1449699929,
+      postMessageResp: {
+        'id_token': tokens.standardIdToken,
+        'access_token': tokens.standardAccessToken,
+        'expires_in': 3600,
+        'token_type': 'Bearer',
+        'state': oauthUtil.mockedState
+      },
+      validateFunc: ({ accessToken, idToken }) => {
+        oauthUtil.validateResponse(accessToken, tokens.standardAccessTokenParsed);
+        oauthUtil.validateResponse(idToken, tokens.standardIdTokenParsed);
+      }
+    });
+  });
+
+  it('should return tokens with authorization server', function() {
+    return oauthUtil.setupFrame({
+      oktaAuthArgs: {
+        pkce: false,
+        issuer: 'https://auth-js-test.okta.com/oauth2/aus8aus76q8iphupD0h7',
+        clientId: 'NPSfOkH5eZrTy8PMDlvx',
+        redirectUri: 'https://example.com/redirect'
+      },
+      tokenRenewTokensArgs: [],
+      postMessageSrc: {
+        baseUri: 'https://auth-js-test.okta.com/oauth2/aus8aus76q8iphupD0h7/v1/authorize',
+        queryParams: {
+          'client_id': 'NPSfOkH5eZrTy8PMDlvx',
+          'redirect_uri': 'https://example.com/redirect',
+          'response_type': 'token id_token',
+          'response_mode': 'okta_post_message',
+          'state': oauthUtil.mockedState,
+          'nonce': oauthUtil.mockedNonce,
+          'scope': 'openid email',
+          'prompt': 'none'
+        }
+      },
+      time: 1449699929,
+      postMessageResp: {
+        'id_token': tokens.authServerIdToken,
+        'access_token': tokens.authServerAccessToken,
+        'expires_in': 3600,
+        'token_type': 'Bearer',
+        'state': oauthUtil.mockedState
+      },
+      validateFunc: (res) => {
+        oauthUtil.validateResponse(res.accessToken, tokens.authServerAccessTokenParsed);
+        oauthUtil.validateResponse(res.idToken, tokens.authServerIdTokenParsed);
+      }
+    });
+  });
+
+  it('should accept tokenParams options', function() {
+    return oauthUtil.setupFrame({
+      oktaAuthArgs: {
+        pkce: false,
+        issuer: 'https://auth-js-test.okta.com',
+        clientId: 'NPSfOkH5eZrTy8PMDlvx',
+        redirectUri: 'https://example.com/redirect'
+      },
+      tokenRenewTokensArgs: [{ scopes: ['openid', 'email', 'profile'] }],
+      postMessageSrc: {
+        baseUri: 'https://auth-js-test.okta.com/oauth2/v1/authorize',
+        queryParams: {
+          'client_id': 'NPSfOkH5eZrTy8PMDlvx',
+          'redirect_uri': 'https://example.com/redirect',
+          'response_type': 'token id_token',
+          'response_mode': 'okta_post_message',
+          'state': oauthUtil.mockedState,
+          'nonce': oauthUtil.mockedNonce,
+          'scope': 'openid email profile',
+          'prompt': 'none'
+        }
+      },
+      time: 1449699929,
+      postMessageResp: {
+        'id_token': tokens.standardIdToken,
+        'access_token': tokens.standardAccessToken,
+        'expires_in': 3600,
+        'token_type': 'Bearer',
+        'state': oauthUtil.mockedState
+      },
+      validateFunc: ({ accessToken, idToken }) => {
+        oauthUtil.validateResponse(accessToken, tokens.standardAccessTokenParsed);
+        oauthUtil.validateResponse(idToken, tokens.standardIdTokenParsed);
+      }
+    });
+  });
+});
+
 describe('token.getUserInfo', function() {
   let responseXHR;
   beforeEach(() => {